Add signMessageBytes for personal_sign call with decoded bytes
signMessageBytes is an alternative to signMessage that only operates on
byte arrays. signMessage calls through to the underlying Ethers
implementation, which interprets all strings as actual strings,
providing no path to signing bytes. Rather than providing a dual
interface with signMessage, we choose to provide a separate call path
that explicitly signs bytes, and leave signMessage to accept strings
that are signed after being decoded to UTF-8 bytes.